Adds WineApi controller to handle API requests
