+++ app/controllers/products_controller.rb
@@ -1,22 +1,10 @@
 class ProductsController < ApplicationController
   before_action :set_product, only: [:show, :edit, :update, :destroy]
 
-  API_URI = 'http://services.wine.com/api/beta2/service.svc/JSON/catalog'.freeze
   # GET /products
   # GET /products.json
   def index
     @products = Product.all
-  end
-
-  # GET /products/request_data
-  def request_data
-    uri = URI(API_URI)
-    query = { :apikey=> '1234' }
-    @response = ::HTTPClient.get(uri, query)
-
-    ::Product.load_api_data(@response.body)
-
-    render :index
   end
 
   # GET /products/1
