+++ app/helpers/wine_api_helper.rb
@@ -1,0 +1,2 @@
+module WineApiHelper
+end
