+++ app/models/product.rb
@@ -17,7 +17,6 @@ class Product
   # VALID_ATTRIBUTES.freeze = ["Id", "Name", "Url", "Type", "Vintage", "Description", "PriceMax", "PriceMin", "PriceRetail"]
 
   def self.load_api_data(data)
-    binding.pry
     product_list = ::JSON.parse(data)["Products"]["List"]
 
     product_list.each do |product|
